Changed the query routing function, moved methods to their respective folder, changed yml for docker
