+++ docker-compose.yml
@@ -12,33 +12,50 @@ services:
     container_name: 'auth'
     ports:
       - 8081:8080
+    depends_on:
+      - mongo
 
   payments:
     build: ./microservices/payments
     container_name: 'payments'
     ports:
       - 8082:8080
+    depends_on:
+      - mongo
 
   products:
     build: ./microservices/products
     container_name: 'products'
     ports:
       - 8083:8080
+    depends_on:
+      - mongo
 
   reports:
     build: ./microservices/reports
     container_name: 'reports'
     ports:
       - 8084:8080
+    depends_on:
+      - mongo
 
   users:
     build: ./microservices/users
     container_name: 'users'
     ports:
       - 8085:8080
+    depends_on:
+      - mongo
 
   gateway:
     build: ./microservices/gateway
     container_name: 'gateway'
     ports:
       - 8080:8080
+    depends_on:
+      - mongo
+      - auth
+      - payments
+      - products
+      - reports
+      - users
