+++ microservices/users/src/methods/get/index.js
@@ -1,6 +1,6 @@
 'use strict'
 
-exports = {
+exports.default = {
   getUsers: require('./users').default,
   getUserById: require('./userById').default
 }
