+++ microservices/users/src/methods/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+exports.default = {
+  ...require('./get/').default,
+  ...require('./post/').default
+}
