+++ microservices/users/src/utils/index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-exports = {
-  emailRegex: require('./utils').emailRegex,
-  passwordRegex: require('./utils').passwordRegex
+exports.default = {
+  ...require('./utils').default
 }
