+++ microservices/users/src/utils/utils.js
@@ -3,7 +3,7 @@
 const emailRegex = /^[a-zA-Z0-9]+([_.-][a-zA-Z0-9]+)*[@][a-z]+[.][a-z]+$/
 const passwordRegex = /^[a-z0-9]{64}$/
 
-exports = {
+exports.default = {
   emailRegex,
   passwordRegex
 }
